some work to do on endpoints in forecast because the scheme isnt following a good design at the moment regards to cragid, id, and stuffs so I cannot tell if population is working

+++ models/models.go
@@ -99,9 +99,9 @@ type DBForecastPayload struct {
 	ScreenTemperature   float64 `json:"screenTemperature"`
 	FeelsLikeTemp       float64 `json:"feelsLikeTemp"`
 	WindSpeed           float64 `json:"windSpeed"`
-	WindDirection       float64 `json:"windDirection"`
+	WindDirection       int     `json:"windDirection"`
 	TotalPrecipAmount   float64 `json:"totalPrecipAmount"`
-	ProbOfPrecipitation float64 `json:"probOfPrecipitation"`
+	ProbOfPrecipitation int     `json:"probOfPrecipitation"`
 	Latitude            float64 `json:"latitude"`
 	Longitude           float64 `json:"longitude"`
 	CragId              int     `json:"cragId"`
